+++ src/main/java/com/autonomouslogic/everef/cli/refdata/esi/EsiTransformer.java
@@ -1,7 +1,0 @@
-package com.autonomouslogic.everef.cli.refdata.esi;
-
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
-public interface EsiTransformer {
-	ObjectNode transformEsi(ObjectNode json, String language) throws Throwable;
-}